fix: date and time are now displayed naturally
Column type in database changed to DATETIME
When displaying dates and times, use `isoformat(' ', 'seconds')`
