+++ .gitignore
@@ -15,3 +15,4 @@
 !run.sh
 !README.md
 !jjit.png
+!/app/classes.py
